Bring across editor changes to edit screen for Android.
